Chore: entity type change #2

+++ backend/moa/src/main/java/com/ssafy/moa/api/entity/Buddy.java
@@ -7,7 +7,7 @@ public class Buddy {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int buddyId;
+    private Long buddyId;
 
     @OneToOne
     @JoinColumns({
